Updates as I worked to solve the ringing of the active peep system
Yet to come is the fix for the assistive ventilation to work with trigger pressures closer to theset PEEP.   Becuase there are escursions above and below that in the first 500-600 mSec of exhalation, I can't set the trigger above ambient atmospheric just yet.   Plan to add  system that watches the measured_pressure and if it stays within a band of 2 above or 2 below the set Peep,  for a run of 10 measurements (approx 30 milliseconds) then re-enable assistive ventilation.